OLS-1715: openai provider allow custom CA
An OpenAI provider can be configured with a custom url. If the
configured endpoint is secured with a self signed certificate, as in the
case of locally hosted OpenAI compatible providers, a trust
configuration needs to be applied.

This PR enables the OpenAI provider to use a custom ca the same way as
the RHELAI and RHOAI Providers.